Move copy page view to wagtail.admin.views.pages.copy
